# unionArrayBy: Find next 1-bits with countTrailingZeros