Closes #374.