@sjakobi 2022-03-24
The benchmark results are pretty surprising: The mean runtime of the `HashMap.union` benchmark has literally doubled from ~60μs to ~120μs!

I think this is due to the weird benchmark data of contiguous `Int`s. Of course it doesn't hurt to increment the index by 1 if the node is full! :/

So, I'll need to improve the benchmark data, maybe simply try with `ByteString`s.

Also, check the generated code.