@sjakobi 2022-03-24
> The benchmark results are pretty surprising: The mean runtime of the HashMap.union benchmark has literally doubled from ~60μs to ~120μs!

These numbers are from my desktop with an i7-4790K CPU.

I'm now using my old laptop with an i3-2350M CPU. For the `union.Int` benchmark I'm seeing a slowdown from 128  μs to 277  μs (116%), for the new `union.ByteString` benchmark I'm also seeing a slowdown from 167  μs to 240  μs (43%).

That's pretty terrible. I'll have to look at the generated code. At least I know that it matters!