@sjakobi 2022-03-25
The latest version (eedc32682e456ec22d5154c4693b0f9058faf4ca) finally brings a speed up for the `union.ByteString` benchmark: 167μs -> 143 μs (-14%).

The `union.Int` benchmark is still slower at 128μs -> 140μs (+9%), but IMHO less realistic.

(Both measurements are from the ancient laptop.)

I should try another version that computes the indices instead of having them as arguments in the loop.

It might also be interesting to combine the various bitmap checks into a single case expression. Maybe GHC can generate something like a lookup table from this?!