@sjakobi 2022-03-26
> I should try another version that computes the indices instead of having them as arguments in the loop.

Much worse, even if I keep `i` as a loop variable.