@sjakobi 2022-03-27
> The latest version ([eedc326](https://github.com/haskell-unordered-containers/unordered-containers/commit/eedc32682e456ec22d5154c4693b0f9058faf4ca)) finally brings a speed up for the `union.ByteString` benchmark: 167μs -> 143 μs (-14%).
> 
> The `union.Int` benchmark is still slower at 128μs -> 140μs (+9%), but IMHO less realistic.
> 
> (Both measurements are from the ancient laptop.)

On my relatively beefier machine with the i7-4790K I'm getting the following numbers:

* `union.Int`: 60μs -> 66μs (+10%)
* `union.ByteString`: 78μs -> 66μs (-15%)