@sjakobi 2022-03-29
> It might also be interesting to combine the various bitmap checks into a single case expression. Maybe GHC can generate something like a lookup table from this?!

I tried this:

```
    let go !i !i1 !i2 !b
            | b == 0 = return ()
            | otherwise = do
                let m = 1 `unsafeShiftL` countTrailingZeros b
                let testBit :: Word -> Word
                    testBit x = fromIntegral (fromEnum (x .&. m /= 0))
                let b'' = b .&. complement m
                let t1 = testBit b1
                let t2 = testBit b2 `unsafeShiftL` 1
                case t1 .|. t2 of
                    1 -> do
                        A.write mary i =<< A.indexM ary1 i1
                        go (i+1) (i1+1)  i2    b''
                    2 -> do
                        A.write mary i =<< A.indexM ary2 i2
                        go (i+1)  i1    (i2+1) b''
                    _ -> do
                        x1 <- A.indexM ary1 i1
                        x2 <- A.indexM ary2 i2
                        A.write mary i $! f x1 x2
                        go (i+1) (i1+1) (i2+1) b''
    go 0 0 0 b'
```

Unfortunately it's slower:

```
    union
      Int:        OK (1.78s)
        162  μs ± 1.6 μs
      ByteString: OK (10.76s)
        156  μs ± 1.5 μs
```