@sjakobi 2022-03-30
I think I'll leave it at this for now. Ultimately it would be nice to recover the perf losses for cases like the `union.Int` benchmark, but I'll leave that for future work.